+++ mkv/ops/salvage.go
@@ -76,7 +76,8 @@ func Salvage(ctx context.Context, srcPath, dstPath string, opts ...mkv.Options) 
 
 	var rb *rollbackBuilder
 	if mkv.RollbackSinkFrom(opts) != nil {
-		rb = newRollbackBuilder()
+		rb = newSpoolingRollbackBuilder(fs, dstPath+".rbspool.tmp")
+		defer rb.cleanup()
 	}
 
 	report, cues, timecodeScale, err := salvageCopy(ctx, srcPath, dstPath, fs, mkv.ProgressFrom(opts), mkv.CleanCutFrom(opts), rb)